refactor(config): split the 3120-line config.go god file by concern
config.go mixed the config schema with the entire load pipeline, the DeepSeek/Mimo pricing tables, and all OS path resolution. Move three cohesive, self-contained concerns into their own files in the same package (pure code motion, no API or behaviour change):

- pricing.go  — DeepSeek/Mimo pricing tables + official-pricing reset/compare helpers
- load.go     — Load/LoadForRoot + the TOML merge and legacy-normalisation pipeline
- paths.go    — user-config / home / session / cache / convention dir resolution

config.go drops from ~3142 to ~1455 lines, leaving the config schema, Default(), and the core resolvers/Validate. The remaining sub-config types stay together as the cohesive config schema — deliberately not split further to avoid fragmenting their cohesion.

Verified: gofmt clean, go vet clean, go build ./..., and go test ./internal/config (English locale). config.go's diff is pure deletion (0 insertions).
