+++ internal/config/config.go
@@ -7,7 +7,6 @@ package config
 
 import (
 	"fmt"
-	"log/slog"
 	"net/netip"
 	"net/url"
 	"os"
@@ -15,8 +14,6 @@ import (
 	"regexp"
 	"runtime"
 	"strings"
-
-	"github.com/BurntSushi/toml"
 
 	"reasonix/internal/fileutil"
 	"reasonix/internal/netclient"
@@ -1209,1693 +1206,6 @@ func Default() *Config {
 			{Name: "deepseek-pro", Kind: "openai", BaseURL: "https://api.deepseek.com", Model: "deepseek-v4-pro", APIKeyEnv: "DEEPSEEK_API_KEY", BalanceURL: "https://api.deepseek.com/user/balance", ContextWindow: 1_000_000, Price: deepSeekV4ProPrice()},
 		},
 	}
-}
-
-func deepSeekV4FlashPrice() *provider.Pricing {
-	return &provider.Pricing{CacheHit: 0.02, Input: 1, Output: 2, Currency: "¥"}
-}
-
-func deepSeekV4ProPrice() *provider.Pricing {
-	return &provider.Pricing{CacheHit: 0.025, Input: 3, Output: 6, Currency: "¥"}
-}
-
-func deepSeekV4Prices() map[string]*provider.Pricing {
-	return map[string]*provider.Pricing{
-		"deepseek-v4-flash": deepSeekV4FlashPrice(),
-		"deepseek-v4-pro":   deepSeekV4ProPrice(),
-	}
-}
-
-func deepSeekV4FlashPriceUSD() *provider.Pricing {
-	return &provider.Pricing{CacheHit: 0.0028, Input: 0.14, Output: 0.28, Currency: "$"}
-}
-
-func deepSeekV4ProPriceUSD() *provider.Pricing {
-	return &provider.Pricing{CacheHit: 0.003625, Input: 0.435, Output: 0.87, Currency: "$"}
-}
-
-func deepSeekV4PricesUSD() map[string]*provider.Pricing {
-	return map[string]*provider.Pricing{
-		"deepseek-v4-flash": deepSeekV4FlashPriceUSD(),
-		"deepseek-v4-pro":   deepSeekV4ProPriceUSD(),
-	}
-}
-
-// DeepSeekV4PricesForLanguage keeps the settings/template call site stable while
-// official DeepSeek defaults move to RMB. Persisted prices still win; this is
-// only used for templates and missing-default backfills.
-func DeepSeekV4PricesForLanguage(lang string) map[string]*provider.Pricing {
-	_ = lang
-	return deepSeekV4Prices()
-}
-
-func deepSeekV4PricesForConfig(c *Config) map[string]*provider.Pricing {
-	_ = c
-	return deepSeekV4Prices()
-}
-
-func deepSeekV4PriceForModel(lang, model string) *provider.Pricing {
-	_ = lang
-	return clonePricing(deepSeekV4Prices()[strings.TrimSpace(model)])
-}
-
-// DeepSeekOfficialPricingLanguage is retained for settings/template compatibility.
-// Official DeepSeek providers now seed RMB prices by default; explicit user
-// prices in config still override these defaults.
-func (c *Config) DeepSeekOfficialPricingLanguage() string {
-	_ = c
-	return "zh"
-}
-
-// ApplyDeepSeekOfficialDefaultPricing refreshes built-in/official DeepSeek
-// prices that still match known official defaults. Custom user prices are left
-// untouched.
-func (c *Config) ApplyDeepSeekOfficialDefaultPricing() {
-	applyDeepSeekOfficialDefaultPricing(c)
-}
-
-func applyDeepSeekOfficialDefaultPricing(c *Config) {
-	if c == nil {
-		return
-	}
-	lang := c.DeepSeekOfficialPricingLanguage()
-	for i := range c.Providers {
-		p := &c.Providers[i]
-		if officialProviderKind(p) != "deepseek" {
-			continue
-		}
-		if isKnownDeepSeekOfficialPricing(p.Model, p.Price) {
-			p.Price = deepSeekV4PriceForModel(lang, p.Model)
-		}
-		for model, price := range p.Prices {
-			if isKnownDeepSeekOfficialPricing(model, price) {
-				p.Prices[model] = deepSeekV4PriceForModel(lang, model)
-			}
-		}
-	}
-}
-
-func mimoV25ProPrice() *provider.Pricing {
-	return &provider.Pricing{CacheHit: 0.025, Input: 3, Output: 6, Currency: "¥"}
-}
-
-func mimoV25Price() *provider.Pricing {
-	return &provider.Pricing{CacheHit: 0.02, Input: 1, Output: 2, Currency: "¥"}
-}
-
-func mimoV2FlashPrice() *provider.Pricing {
-	return &provider.Pricing{CacheHit: 0.07, Input: 0.70, Output: 2.10, Currency: "¥"}
-}
-
-func mimoDomesticPrices(models []string) map[string]*provider.Pricing {
-	prices := map[string]*provider.Pricing{}
-	for _, model := range models {
-		switch strings.TrimSpace(model) {
-		case "mimo-v2.5-pro", "mimo-v2-pro":
-			prices[model] = mimoV25ProPrice()
-		case "mimo-v2.5", "mimo-v2-omni":
-			prices[model] = mimoV25Price()
-		case "mimo-v2-flash":
-			prices[model] = mimoV2FlashPrice()
-		}
-	}
-	return prices
-}
-
-// ResetOfficialProviderPricingOnUpgrade resets official DeepSeek prices to
-// the current built-in RMB defaults once for desktop upgrades. It intentionally
-// runs from the desktop app startup path, not every config Load(), so user edits
-// made after the upgrade are preserved.
-func ResetOfficialProviderPricingOnUpgrade(path string) (bool, error) {
-	path = strings.TrimSpace(path)
-	if path == "" {
-		return false, nil
-	}
-	if _, err := os.Stat(path); err != nil {
-		if os.IsNotExist(err) {
-			return false, nil
-		}
-		return false, err
-	}
-	var header Config
-	if _, err := toml.DecodeFile(path, &header); err != nil {
-		return false, fmt.Errorf("config %s: %w", path, err)
-	}
-	if header.ConfigVersion >= Default().ConfigVersion {
-		return false, nil
-	}
-	cfg := LoadForEdit(path)
-	resetOfficialProviderPricingDefaults(cfg)
-	cfg.ConfigVersion = Default().ConfigVersion
-	if err := cfg.SaveTo(path); err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
-func resetOfficialProviderPricingDefaults(c *Config) {
-	if c == nil {
-		return
-	}
-	for i := range c.Providers {
-		p := &c.Providers[i]
-		switch {
-		case officialProviderKind(p) == "deepseek":
-			resetDeepSeekOfficialPricing(p)
-		}
-	}
-}
-
-func resetDeepSeekOfficialPricing(p *ProviderEntry) {
-	if p == nil {
-		return
-	}
-	defaults := deepSeekV4Prices()
-	p.Price = nil
-	if strings.TrimSpace(p.Model) != "" && len(p.Models) == 0 {
-		if price := defaults[strings.TrimSpace(p.Model)]; price != nil {
-			p.Price = clonePricing(price)
-			p.Prices = nil
-			return
-		}
-	}
-	if p.Prices == nil {
-		p.Prices = map[string]*provider.Pricing{}
-	}
-	for model, price := range defaults {
-		if p.HasModel(model) {
-			p.Prices[model] = clonePricing(price)
-		}
-	}
-}
-
-func isKnownDeepSeekOfficialPricing(model string, price *provider.Pricing) bool {
-	model = strings.TrimSpace(model)
-	if model == "" || price == nil {
-		return false
-	}
-	for _, prices := range []map[string]*provider.Pricing{deepSeekV4Prices(), deepSeekV4PricesUSD()} {
-		if samePricing(price, prices[model]) {
-			return true
-		}
-	}
-	return false
-}
-
-func samePricing(a, b *provider.Pricing) bool {
-	if a == nil || b == nil {
-		return false
-	}
-	return a.CacheHit == b.CacheHit && a.Input == b.Input && a.Output == b.Output && a.Currency == b.Currency
-}
-
-// Load builds the configuration: defaults, then user config, then project
-// config, then MCP servers from Claude Code's .mcp.json, then (lowest priority)
-// the v0.x ~/.reasonix/config.json's mcpServers. Provider api_key_env values
-// resolve from Reasonix's global .env, not from project .env files.
-func Load() (*Config, error) {
-	return LoadForRoot(".")
-}
-
-// LoadForRoot builds the configuration with project files resolved from root
-// instead of the current working directory. When root is "" or ".", it behaves
-// like Load(). This is the workspace-aware entry point: desktop tabs use it so
-// each project's reasonix.toml + .mcp.json are resolved independently without
-// changing the process cwd, while provider keys stay rooted in Reasonix home.
-func LoadForRoot(root string) (*Config, error) {
-	root = resolveRoot(root)
-	expansionEnv := loadDotEnvForRoot(root)
-	cfg := Default()
-	cfg.setExpansionEnv(expansionEnv)
-	cfg.CredentialsStore = credentialsStoreMode()
-
-	projectTOML := "reasonix.toml"
-	if root != "." {
-		projectTOML = filepath.Join(root, "reasonix.toml")
-	}
-
-	var tomlSources []string
-	if uc := userConfigLoadPath(); uc != "" {
-		tomlSources = append(tomlSources, uc)
-		if err := mergeRuntimeTOMLFile(cfg, uc); err != nil {
-			return nil, err
-		}
-	}
-	globalMaxSteps := cfg.Agent.MaxSteps
-	globalPlannerMaxSteps := cfg.Agent.PlannerMaxSteps
-
-	tomlSources = append(tomlSources, projectTOML)
-	if err := mergeRuntimeTOMLFile(cfg, projectTOML); err != nil {
-		return nil, err
-	}
-	// Runtime step caps are user/global controls, not project policy. Keep the
-	// project config's other fields, but do not let ./reasonix.toml override
-	// the user's execution and planner round limits.
-	cfg.Agent.MaxSteps = globalMaxSteps
-	cfg.Agent.PlannerMaxSteps = globalPlannerMaxSteps
-	// toml.DecodeFile replaces [[plugins]] wholesale, so cfg.Plugins now holds
-	// only the last file's. Re-merge by name across all sources (later wins) so a
-	// project reasonix.toml doesn't drop the global config's MCP servers.
-	plugins, err := mergeTOMLPlugins(tomlSources)
-	if err != nil {
-		return nil, err
-	}
-	cfg.Plugins = plugins
-	if providers, providerSources, shadowedProjectProviders, ok, err := mergeTOMLProviders(tomlSources); err != nil {
-		return nil, err
-	} else if ok {
-		cfg.Providers = providers
-		cfg.providerSources = providerSources
-		cfg.shadowedProjectProviders = shadowedProjectProviders
-	}
-	if access, ok, err := mergeTOMLProviderAccess(tomlSources); err != nil {
-		return nil, err
-	} else if ok {
-		cfg.Desktop.ProviderAccess = access
-	}
-
-	// Claude Code's .mcp.json (project root) is read last and merged into
-	// [[plugins]], so a server configured for Claude works here unchanged.
-	// reasonix.toml wins on a name collision (see mergeMCPJSON).
-	mcpFile := mcpJSONFile
-	if root != "." {
-		mcpFile = filepath.Join(root, mcpJSONFile)
-	}
-	entries, err := loadMCPJSON(mcpFile)
-	if err != nil {
-		return nil, err
-	}
-	cfg.mergeMCPJSON(entries)
-
-	// Lowest priority: the v0.x ~/.reasonix/config.json's mcpServers, so upgrading
-	// from the TypeScript line keeps MCP servers without rewriting them. Anything
-	// the v2 config or .mcp.json already declared wins on a name collision.
-	cfg.mergeMCPJSON(loadLegacyMCP(legacyConfigPath()))
-	normalizePluginCommandLines(cfg)
-	normalizeLegacyEffort(cfg)
-	normalizeLegacyMCPTiers(cfg)
-	normalizeLegacyMimoCustomProviders(cfg)
-	normalizeLegacyProviderModels(cfg)
-	normalizeDesktopOfficialProviderAccess(cfg)
-	normalizeOfficialDeepSeekModels(cfg)
-	applyDeepSeekOfficialDefaultPricing(cfg)
-	backfillDeepSeekOfficialPrices(cfg)
-	normalizeEffortConfig(cfg)
-	backfillDeepSeekPro(cfg)
-	cfg.Agent.AutoPlan = userAutoPlanMode()
-	cfg.CredentialsStore = credentialsStoreMode()
-	cfg.setExpansionEnv(expansionEnv)
-	resolveProviderCredentialsForRoot(root, cfg)
-	return cfg, nil
-}
-
-func (c *Config) setExpansionEnv(env map[string]string) {
-	if c == nil {
-		return
-	}
-	c.expansionEnv = cloneStringMap(env)
-	for i := range c.Plugins {
-		c.Plugins[i].expansionEnv = c.expansionEnv
-	}
-}
-
-func cloneStringMap(in map[string]string) map[string]string {
-	if len(in) == 0 {
-		return nil
-	}
-	out := make(map[string]string, len(in))
-	for k, v := range in {
-		out[k] = v
-	}
-	return out
-}
-
-func userAutoPlanMode() string {
-	cfg := Default()
-	if uc := userConfigLoadPath(); uc != "" {
-		_ = mergeFile(cfg, uc)
-	}
-	switch strings.ToLower(strings.TrimSpace(cfg.Agent.AutoPlan)) {
-	case "on", "ask":
-		return "on"
-	default:
-		return "off"
-	}
-}
-
-// backfillDeepSeekPro restores deepseek-pro for configs the pre-fix setup wizard
-// wrote with only deepseek-v4-flash: a keyless /models probe used to drop the Pro
-// SKU, leaving users unable to switch to it. In-memory only — the user's file is
-// untouched. Narrowly scoped to the official DeepSeek endpoint (which is known to
-// serve pro) so a custom flash-only deployment isn't given an entry that 404s.
-func backfillDeepSeekPro(c *Config) {
-	const flashModel, proModel = "deepseek-v4-flash", "deepseek-v4-pro"
-	var flash *ProviderEntry
-	for i := range c.Providers {
-		p := &c.Providers[i]
-		if p.Name == "deepseek-pro" {
-			return
-		}
-		for _, m := range p.ModelList() {
-			switch m {
-			case proModel:
-				return // pro already reachable
-			case flashModel:
-				if strings.Contains(p.BaseURL, "api.deepseek.com") {
-					flash = p
-				}
-			}
-		}
-	}
-	if flash == nil {
-		return
-	}
-	// If the user has explicitly curated a model list for the flash provider
-	// (e.g. unchecked pro in Settings), respect that choice and do not backfill.
-	if len(flash.Models) > 0 {
-		return
-	}
-	for _, bp := range Default().Providers {
-		if bp.Name == "deepseek-pro" {
-			bp.APIKeyEnv = flash.APIKeyEnv
-			bp.Price = deepSeekV4PriceForModel(c.DeepSeekOfficialPricingLanguage(), proModel)
-			c.Providers = append(c.Providers, bp)
-			return
-		}
-	}
-}
-
-func backfillDeepSeekOfficialPrices(c *Config) {
-	if c == nil {
-		return
-	}
-	defaults := deepSeekV4PricesForConfig(c)
-	for i := range c.Providers {
-		p := &c.Providers[i]
-		if officialProviderKind(p) != "deepseek" {
-			continue
-		}
-		if p.Price != nil {
-			continue
-		}
-		if p.Prices == nil {
-			p.Prices = map[string]*provider.Pricing{}
-		}
-		for model, price := range defaults {
-			if p.HasModel(model) && p.Prices[model] == nil {
-				p.Prices[model] = clonePricing(price)
-			}
-		}
-	}
-}
-
-func officialProviderKind(p *ProviderEntry) string {
-	if p == nil {
-		return ""
-	}
-	u, err := url.Parse(strings.TrimSpace(p.BaseURL))
-	if err != nil {
-		return ""
-	}
-	if strings.EqualFold(u.Hostname(), "api.deepseek.com") {
-		return "deepseek"
-	}
-	return ""
-}
-
-func resolveRoot(root string) string {
-	if root == "" || root == "." {
-		return "."
-	}
-	return filepath.Clean(root)
-}
-
-// normalizeLegacyEffort migrates the retired DeepSeek effort="off" (the old
-// /thinking off that disabled thinking) to the provider default, so a config
-// written by an older version keeps loading instead of erroring on a value the
-// provider no longer accepts.
-func normalizeLegacyEffort(c *Config) {
-	for i := range c.Providers {
-		if strings.EqualFold(strings.TrimSpace(c.Providers[i].Effort), "off") {
-			c.Providers[i].Effort = ""
-		}
-	}
-}
-
-// mergeTOMLPlugins merges [[plugins]] across TOML sources by name (later source wins).
-func mergeTOMLPlugins(paths []string) ([]PluginEntry, error) {
-	var merged []PluginEntry
-	index := map[string]int{}
-	for _, path := range paths {
-		if _, err := os.Stat(path); err != nil {
-			continue
-		}
-		var f Config
-		if _, err := toml.DecodeFile(path, &f); err != nil {
-			return nil, fmt.Errorf("config %s: %w", path, err)
-		}
-		for _, p := range f.Plugins {
-			p, _ = NormalizePluginCommandLine(p)
-			if i, ok := index[p.Name]; ok {
-				merged[i] = p
-				continue
-			}
-			index[p.Name] = len(merged)
-			merged = append(merged, p)
-		}
-	}
-	return merged, nil
-}
-
-// mergeTOMLProviders merges [[providers]] across TOML sources by provider name.
-// User-global providers win over same-named project providers; project providers
-// only fill names the global config does not define. Keep official legacy aliases
-// distinct here: they can carry different default models and effort capabilities,
-// and the later desktop normalization layer handles canonical Settings access.
-func mergeTOMLProviders(paths []string) ([]ProviderEntry, map[string]providerSourceScope, []ProviderEntry, bool, error) {
-	var merged []ProviderEntry
-	var shadowedProject []ProviderEntry
-	index := map[string]int{}
-	sources := map[string]providerSourceScope{}
-	saw := false
-	for _, path := range paths {
-		if _, err := os.Stat(path); err != nil {
-			continue
-		}
-		var f Config
-		if _, err := toml.DecodeFile(path, &f); err != nil {
-			return nil, nil, nil, false, fmt.Errorf("config %s: %w", path, err)
-		}
-		if len(f.Providers) == 0 {
-			continue
-		}
-		saw = true
-		source := providerSourceForPath(path)
-		for _, p := range f.Providers {
-			normalizeProviderEffortFields(&p)
-			key := providerMergeKey(p)
-			if i, ok := index[key]; ok {
-				if sources[key] == providerSourceProject && source == providerSourceUser {
-					shadowedProject = append(shadowedProject, merged[i])
-					merged[i] = p
-					sources[key] = source
-				} else if sources[key] == providerSourceUser && source == providerSourceProject {
-					shadowedProject = append(shadowedProject, p)
-				}
-				continue
-			} else {
-				index[key] = len(merged)
-				merged = append(merged, p)
-				sources[key] = source
-			}
-		}
-	}
-	return merged, sources, shadowedProject, saw, nil
-}
-
-func providerSourceForPath(path string) providerSourceScope {
-	if isUserConfigPath(path) {
-		return providerSourceUser
-	}
-	return providerSourceProject
-}
-
-func providerMergeKey(p ProviderEntry) string {
-	return strings.TrimSpace(p.Name)
-}
-
-// mergeTOMLProviderAccess merges desktop.provider_access across TOML sources so
-// project desktop settings do not hide account-level providers from the desktop
-// model switcher.
-func mergeTOMLProviderAccess(paths []string) ([]string, bool, error) {
-	var merged []string
-	seen := map[string]bool{}
-	saw := false
-	for _, path := range paths {
-		if _, err := os.Stat(path); err != nil {
-			continue
-		}
-		var f Config
-		meta, err := toml.DecodeFile(path, &f)
-		if err != nil {
-			return nil, false, fmt.Errorf("config %s: %w", path, err)
-		}
-		if !meta.IsDefined("desktop", "provider_access") {
-			continue
-		}
-		saw = true
-		for _, name := range f.Desktop.ProviderAccess {
-			name = strings.TrimSpace(name)
-			if name == "" || seen[name] {
-				continue
-			}
-			seen[name] = true
-			merged = append(merged, name)
-		}
-	}
-	return merged, saw, nil
-}
-
-// LoadForEdit returns a config to seed the `reasonix setup` wizard when reconfiguring:
-// the built-in defaults with the file at path (if present) decoded on top, so a
-// reconfigure preserves the user's existing providers and agent settings instead
-// of resetting to defaults. Reasonix's global .env is loaded so api_key_env
-// resolution works while the wizard decides which keys are still missing.
-func LoadForEdit(path string) *Config {
-	cfg, err := loadForEditStrict(path, true)
-	if err == nil {
-		return cfg
-	}
-	slog.Warn("config: load for edit failed, using defaults", "path", path, "err", err)
-	loadDotEnvForEditPath(path)
-	cfg = Default()
-	normalizeConfigForEdit(cfg)
-	return cfg
-}
-
-func LoadForEditWithoutCredentials(path string) *Config {
-	cfg, err := loadForEditStrict(path, false)
-	if err == nil {
-		return cfg
-	}
-	slog.Warn("config: load for edit failed, using defaults", "path", path, "err", err)
-	cfg = Default()
-	normalizeConfigForEdit(cfg)
-	return cfg
-}
-
-func loadForEditStrict(path string, loadCredentials bool) (*Config, error) {
-	if loadCredentials {
-		loadDotEnvForEditPath(path)
-	}
-	cfg := Default()
-	if _, err := os.Stat(path); err == nil {
-		if err := migrateLegacyMCPTiersFile(path); err != nil {
-			return nil, fmt.Errorf("config %s: %w", path, err)
-		}
-	}
-	if err := mergeFile(cfg, path); err != nil {
-		return nil, err
-	}
-	migratedMimo := normalizeConfigForEdit(cfg)
-	if migratedMimo && strings.TrimSpace(path) != "" {
-		if _, err := os.Stat(path); err == nil {
-			if err := cfg.SaveTo(path); err != nil {
-				return nil, err
-			}
-		}
-	}
-	return cfg, nil
-}
-
-func normalizeConfigForEdit(cfg *Config) bool {
-	normalizePluginCommandLines(cfg)
-	normalizeLegacyEffort(cfg)
-	normalizeLegacyMCPTiers(cfg)
-	migratedMimo := normalizeLegacyMimoCustomProviders(cfg)
-	normalizeLegacyProviderModels(cfg)
-	normalizeDesktopOfficialProviderAccess(cfg)
-	applyDeepSeekOfficialDefaultPricing(cfg)
-	backfillDeepSeekOfficialPrices(cfg)
-	normalizeEffortConfig(cfg)
-	return migratedMimo
-}
-
-func loadDotEnvForEditPath(path string) {
-	path = strings.TrimSpace(path)
-	if path == "" || isUserConfigPath(path) {
-		loadDotEnv()
-		return
-	}
-	loadDotEnvForRoot(filepath.Dir(path))
-}
-
-// mergeFile decodes a TOML file onto cfg if it exists. An absent file is not an error.
-func mergeFile(cfg *Config, path string) error {
-	if _, err := os.Stat(path); err != nil {
-		return nil
-	}
-	if _, err := toml.DecodeFile(path, cfg); err != nil {
-		return fmt.Errorf("config %s: %w", path, err)
-	}
-	return nil
-}
-
-func mergeRuntimeTOMLFile(cfg *Config, path string) error {
-	if _, err := os.Stat(path); err == nil {
-		if err := migrateLegacyMCPTiersFile(path); err != nil {
-			slog.Warn("config: legacy mcp tier migration failed", "path", path, "err", err)
-		}
-	}
-	return mergeFile(cfg, path)
-}
-
-// normalizeLegacyMCPTiers keeps loaded legacy config files on the new product
-// behavior: enabled MCP servers connect in the background by default, and the
-// retired per-server startup tier is no longer a user-facing setting.
-func normalizeLegacyMCPTiers(c *Config) {
-	if c == nil {
-		return
-	}
-	for i := range c.Plugins {
-		c.Plugins[i].Tier = ""
-	}
-}
-
-func migrateLegacyMCPTiersFile(path string) error {
-	info, err := os.Stat(path)
-	if err != nil {
-		return err
-	}
-	raw, err := os.ReadFile(path)
-	if err != nil {
-		return err
-	}
-	next, changed := stripLegacyMCPTierLines(string(raw))
-	if !changed {
-		return nil
-	}
-	return os.WriteFile(path, []byte(next), info.Mode().Perm())
-}
-
-func stripLegacyMCPTierLines(raw string) (string, bool) {
-	lines := strings.Split(raw, "\n")
-	section := ""
-	changed := false
-	out := make([]string, 0, len(lines))
-	for _, line := range lines {
-		if header := tomlSectionHeader(line); header != "" {
-			section = header
-		}
-		if section == "plugins" && isTOMLKeyAssignment(line, "tier") {
-			changed = true
-			continue
-		}
-		out = append(out, line)
-	}
-	return strings.Join(out, "\n"), changed
-}
-
-func tomlSectionHeader(line string) string {
-	trimmed := strings.TrimSpace(line)
-	if !strings.HasPrefix(trimmed, "[") {
-		return ""
-	}
-	if i := strings.Index(trimmed, "#"); i >= 0 {
-		trimmed = strings.TrimSpace(trimmed[:i])
-	}
-	switch trimmed {
-	case "[[plugins]]":
-		return "plugins"
-	default:
-		return "other"
-	}
-}
-
-func isTOMLKeyAssignment(line, key string) bool {
-	trimmed := strings.TrimSpace(line)
-	if strings.HasPrefix(trimmed, "#") || !strings.HasPrefix(trimmed, key) {
-		return false
-	}
-	rest := strings.TrimSpace(strings.TrimPrefix(trimmed, key))
-	return strings.HasPrefix(rest, "=")
-}
-
-// normalizeLegacyProviderModels repairs provider entries written by older
-// desktop builds that carried the official provider name/endpoint but omitted the
-// model field. The repair is intentionally narrow: valid user-provided model
-// lists are left untouched, while known official aliases get the model implied by
-// their preset name so model pickers and provider validation have an option.
-func normalizeLegacyProviderModels(c *Config) {
-	if c == nil {
-		return
-	}
-	for i := range c.Providers {
-		p := &c.Providers[i]
-		if providerHasAnyModel(*p) {
-			continue
-		}
-		if model := legacyOfficialProviderModel(p.Name); model != "" {
-			p.Model = model
-		}
-	}
-}
-
-func normalizeLegacyMimoProviderCatalogs(c *Config) bool {
-	if c == nil {
-		return false
-	}
-	changed := false
-	for i := range c.Providers {
-		p := &c.Providers[i]
-		if legacyMimoProviderName(p.Name) == "" || len(p.Models) > 0 {
-			continue
-		}
-		switch officialProviderHost(p.BaseURL) {
-		case "api.xiaomimimo.com":
-			if applyLegacyMimoCatalog(p, legacyMimoAPIModels(), []string{"mimo-v2.5", "mimo-v2-omni"}, "mimo-v2.5-pro") {
-				changed = true
-			}
-		case "token-plan-cn.xiaomimimo.com":
-			if applyLegacyMimoCatalog(p, legacyMimoTokenPlanModels(), []string{"mimo-v2.5"}, "mimo-v2.5-pro") {
-				changed = true
-			}
-		}
-	}
-	return changed
-}
-
-func applyLegacyMimoCatalog(p *ProviderEntry, models, visionModels []string, fallbackDefault string) bool {
-	if p == nil || len(models) == 0 {
-		return false
-	}
-	beforeModels := append([]string(nil), p.Models...)
-	beforeVision := append([]string(nil), p.VisionModels...)
-	beforeDefault := p.Default
-	beforeModel := p.Model
-	beforeWindow := p.ContextWindow
-	beforeNoProxy := p.NoProxy
-	beforePricesLen := len(p.Prices)
-
-	currentDefault := strings.TrimSpace(p.Default)
-	if currentDefault == "" {
-		currentDefault = strings.TrimSpace(p.Model)
-	}
-	p.Models = mergeModelLists(models, p.ModelList())
-	p.Model = p.Models[0]
-	p.Default = firstKnownModel(currentDefault, p.Models, fallbackDefault)
-	p.VisionModels = mergeModelLists(visionModels, p.VisionModels)
-	backfillOfficialContextWindow(p, 1_048_576)
-	p.NoProxy = true
-	if p.Prices == nil {
-		p.Prices = mimoDomesticPrices(models)
-	} else {
-		for model, price := range mimoDomesticPrices(models) {
-			if p.Prices[model] == nil {
-				p.Prices[model] = price
-			}
-		}
-	}
-
-	return !stringSlicesEqual(beforeModels, p.Models) ||
-		!stringSlicesEqual(beforeVision, p.VisionModels) ||
-		beforeDefault != p.Default ||
-		beforeModel != p.Model ||
-		beforeWindow != p.ContextWindow ||
-		beforeNoProxy != p.NoProxy ||
-		beforePricesLen != len(p.Prices)
-}
-
-func stringSlicesEqual(a, b []string) bool {
-	if len(a) != len(b) {
-		return false
-	}
-	for i := range a {
-		if a[i] != b[i] {
-			return false
-		}
-	}
-	return true
-}
-
-func normalizeOfficialDeepSeekModels(c *Config) {
-	if c == nil {
-		return
-	}
-	for i := range c.Providers {
-		p := &c.Providers[i]
-		if officialProviderHost(p.BaseURL) != "api.deepseek.com" {
-			continue
-		}
-		switch strings.TrimSpace(p.Name) {
-		case "deepseek":
-			ensureProviderModels(p, []string{"deepseek-v4-flash", "deepseek-v4-pro"}, "deepseek-v4-flash")
-		case "deepseek-flash":
-			ensureProviderModels(p, []string{"deepseek-v4-flash"}, "deepseek-v4-flash")
-		case "deepseek-pro":
-			ensureProviderModels(p, []string{"deepseek-v4-pro"}, "deepseek-v4-pro")
-		}
-	}
-}
-
-func officialProviderHost(baseURL string) string {
-	u, err := url.Parse(strings.TrimSpace(baseURL))
-	if err != nil {
-		return ""
-	}
-	return strings.ToLower(u.Hostname())
-}
-
-func ensureProviderModels(p *ProviderEntry, required []string, fallbackDefault string) {
-	if p == nil {
-		return
-	}
-	// If the user has explicitly curated a model list (via Settings), respect
-	// that choice and do not merge additional required models.
-	if len(p.Models) > 0 {
-		return
-	}
-	models := mergeModelLists(required, p.ModelList())
-	if len(models) == 0 {
-		return
-	}
-	p.Model = models[0]
-	if len(models) > 1 {
-		p.Models = models
-		p.Default = firstKnownModel(p.Default, models, fallbackDefault)
-		return
-	}
-	p.Models = nil
-	p.Default = ""
-}
-
-func legacyOfficialProviderModel(name string) string {
-	switch strings.TrimSpace(name) {
-	case "deepseek-flash":
-		return "deepseek-v4-flash"
-	case "deepseek-pro":
-		return "deepseek-v4-pro"
-	case "mimo", "xiaomi-mimo", "xiaomi_mimo", "mimo-api", "mimo-token-plan", "mimo-pro":
-		return "mimo-v2.5-pro"
-	case "mimo-flash":
-		return "mimo-v2.5"
-	default:
-		return ""
-	}
-}
-
-func normalizeLegacyMimoCustomProviders(c *Config) bool {
-	return normalizeLegacyMimoCustomProvidersForRefs(c, legacyMimoConfigRefs(c)...)
-}
-
-// NormalizeLegacyMimoCustomProvidersForRefs appends custom OpenAI-compatible
-// MiMo providers needed by legacy refs that live outside reasonix.toml, such as
-// restored desktop tab state.
-func NormalizeLegacyMimoCustomProvidersForRefs(c *Config, refs ...string) bool {
-	return normalizeLegacyMimoCustomProvidersForRefs(c, refs...)
-}
-
-func normalizeLegacyMimoCustomProvidersForRefs(c *Config, refs ...string) bool {
-	if c == nil {
-		return false
-	}
-	needed := map[string]bool{}
-	addRef := func(ref string) {
-		if name := legacyMimoProviderNameForRef(ref); name != "" {
-			needed[name] = true
-		}
-	}
-	for _, ref := range refs {
-		addRef(ref)
-	}
-	changed := normalizeLegacyMimoProviderCatalogs(c)
-	for name := range needed {
-		if _, ok := c.Provider(name); ok {
-			continue
-		}
-		c.Providers = append(c.Providers, legacyMimoCustomProvider(name))
-		changed = true
-	}
-	if normalizeLegacyMimoProviderCatalogs(c) {
-		changed = true
-	}
-	return changed
-}
-
-func legacyMimoConfigRefs(c *Config) []string {
-	if c == nil {
-		return nil
-	}
-	refs := []string{
-		c.DefaultModel,
-		c.Agent.PlannerModel,
-		c.Agent.SubagentModel,
-		c.Agent.AutoPlanClassifier,
-		c.Bot.Model,
-	}
-	for _, ref := range c.Agent.SubagentModels {
-		refs = append(refs, ref)
-	}
-	for _, conn := range c.Bot.Connections {
-		refs = append(refs, conn.Model)
-	}
-	refs = append(refs, c.Desktop.ProviderAccess...)
-	return refs
-}
-
-func legacyMimoProviderName(ref string) string {
-	switch strings.TrimSpace(ref) {
-	case "mimo", "xiaomi-mimo", "xiaomi_mimo", "mimo-api", "mimo-token-plan", "mimo-pro", "mimo-flash":
-		return strings.TrimSpace(ref)
-	default:
-		return ""
-	}
-}
-
-func legacyMimoProviderNameForRef(ref string) string {
-	ref = strings.TrimSpace(ref)
-	if ref == "" {
-		return ""
-	}
-	providerName, _, hasModel := strings.Cut(ref, "/")
-	if name := legacyMimoProviderName(providerName); name != "" {
-		return name
-	}
-	if hasModel {
-		return ""
-	}
-	switch ref {
-	case "mimo-v2.5-pro":
-		return "mimo-pro"
-	case "mimo-v2.5":
-		return "mimo-flash"
-	case "mimo-v2-omni":
-		return "mimo-api"
-	default:
-		return ""
-	}
-}
-
-func legacyMimoAPIModels() []string {
-	return []string{"mimo-v2.5-pro", "mimo-v2.5", "mimo-v2-omni"}
-}
-
-func legacyMimoTokenPlanModels() []string {
-	return []string{"mimo-v2.5-pro", "mimo-v2.5"}
-}
-
-func legacyMimoCustomProvider(name string) ProviderEntry {
-	switch strings.TrimSpace(name) {
-	case "mimo", "xiaomi-mimo", "xiaomi_mimo", "mimo-api":
-		models := legacyMimoAPIModels()
-		return ProviderEntry{
-			Name:          strings.TrimSpace(name),
-			Kind:          "openai",
-			BaseURL:       "https://api.xiaomimimo.com/v1",
-			Models:        models,
-			VisionModels:  []string{"mimo-v2.5", "mimo-v2-omni"},
-			Default:       "mimo-v2.5-pro",
-			APIKeyEnv:     "MIMO_API_KEY",
-			ContextWindow: 1_048_576,
-			Prices:        mimoDomesticPrices(models),
-			NoProxy:       true,
-		}
-	case "mimo-token-plan":
-		models := legacyMimoTokenPlanModels()
-		return ProviderEntry{
-			Name:          "mimo-token-plan",
-			Kind:          "openai",
-			BaseURL:       "https://token-plan-cn.xiaomimimo.com/v1",
-			Models:        models,
-			VisionModels:  []string{"mimo-v2.5"},
-			Default:       "mimo-v2.5-pro",
-			APIKeyEnv:     "MIMO_API_KEY",
-			ContextWindow: 1_048_576,
-			Prices:        mimoDomesticPrices(models),
-			NoProxy:       true,
-		}
-	case "mimo-flash":
-		return ProviderEntry{Name: "mimo-flash", Kind: "openai", BaseURL: "https://token-plan-cn.xiaomimimo.com/v1", Model: "mimo-v2.5", APIKeyEnv: "MIMO_API_KEY", ContextWindow: 1_000_000, Price: mimoV25Price(), NoProxy: true}
-	default:
-		return ProviderEntry{Name: "mimo-pro", Kind: "openai", BaseURL: "https://token-plan-cn.xiaomimimo.com/v1", Model: "mimo-v2.5-pro", APIKeyEnv: "MIMO_API_KEY", ContextWindow: 1_000_000, Price: mimoV25ProPrice(), NoProxy: true}
-	}
-}
-
-func normalizeDesktopOfficialProviderAccess(c *Config) {
-	if c == nil || len(c.Desktop.ProviderAccess) == 0 {
-		return
-	}
-	seen := desktopProviderAccessMap(nil)
-	next := make([]string, 0, len(c.Desktop.ProviderAccess))
-	for _, name := range c.Desktop.ProviderAccess {
-		name = desktopProviderAccessNameForConfig(c, name)
-		if name == "" || seen[name] {
-			continue
-		}
-		seen[name] = true
-		next = append(next, name)
-	}
-	c.Desktop.ProviderAccess = next
-	if seen["deepseek"] {
-		ensureDeepSeekOfficialProvider(c)
-	}
-	normalizeLegacyMimoProviderCatalogs(c)
-	retargetDesktopOfficialRefs(c, seen)
-}
-
-// NormalizeLegacyDesktopProviderAccess seeds the desktop provider-access list
-// for configs written before Settings tracked explicit provider access. Callers
-// should only use this when they know the TOML did not declare provider_access;
-// an explicit empty list means the user removed all access entries.
-func NormalizeLegacyDesktopProviderAccess(c *Config) {
-	if c == nil || len(c.Desktop.ProviderAccess) > 0 {
-		return
-	}
-	seen := desktopProviderAccessMap(nil)
-	var access []string
-	add := func(name string) {
-		name = desktopProviderAccessNameForConfig(c, name)
-		if name == "" || seen[name] {
-			return
-		}
-		seen[name] = true
-		access = append(access, name)
-	}
-	addRef := func(ref string) {
-		if entry, ok := c.ResolveModel(ref); ok {
-			if !entry.Configured() {
-				return
-			}
-			add(entry.Name)
-		}
-	}
-	addRef(c.DefaultModel)
-	addRef(c.Agent.PlannerModel)
-	addRef(c.Agent.SubagentModel)
-	addRef(c.Agent.AutoPlanClassifier)
-	for _, ref := range c.Agent.SubagentModels {
-		addRef(ref)
-	}
-	addRef(c.Bot.Model)
-	for _, conn := range c.Bot.Connections {
-		addRef(conn.Model)
-	}
-	for i := range c.Providers {
-		p := &c.Providers[i]
-		if legacyMimoProviderName(p.Name) != "" && len(p.ModelList()) > 0 {
-			add(p.Name)
-			continue
-		}
-		if p.Configured() && len(p.ModelList()) > 0 {
-			add(p.Name)
-		}
-	}
-	if len(access) == 0 {
-		return
-	}
-	c.Desktop.ProviderAccess = access
-	normalizeDesktopOfficialProviderAccess(c)
-}
-
-func canonicalDesktopOfficialProviderName(name string) string {
-	switch strings.TrimSpace(name) {
-	case "deepseek-flash", "deepseek-pro":
-		return "deepseek"
-	default:
-		return strings.TrimSpace(name)
-	}
-}
-
-func desktopProviderAccessNameForConfig(c *Config, name string) string {
-	name = strings.TrimSpace(name)
-	if name == "" {
-		return ""
-	}
-	canonical := canonicalDesktopOfficialProviderName(name)
-	if canonical == name {
-		return name
-	}
-	if c == nil {
-		return canonical
-	}
-	if p, ok := c.Provider(name); ok && !providerEntryMatchesCanonicalOfficialAccess(p, canonical) {
-		return name
-	}
-	return canonical
-}
-
-func providerEntryMatchesCanonicalOfficialAccess(p *ProviderEntry, canonical string) bool {
-	if p == nil {
-		return false
-	}
-	switch canonical {
-	case "deepseek":
-		return officialProviderKind(p) == "deepseek"
-	default:
-		return false
-	}
-}
-
-// CanonicalDesktopOfficialProviderName returns the Settings Center provider ID
-// for built-in official provider aliases.
-func CanonicalDesktopOfficialProviderName(name string) string {
-	return canonicalDesktopOfficialProviderName(name)
-}
-
-func desktopProviderAccessMap(names []string) map[string]bool {
-	out := map[string]bool{}
-	for _, name := range names {
-		name = strings.TrimSpace(name)
-		if name != "" {
-			out[name] = true
-		}
-	}
-	return out
-}
-
-func ensureDeepSeekOfficialProvider(c *Config) {
-	if p, ok := c.Provider("deepseek"); ok {
-		if officialProviderKind(p) == "deepseek" {
-			backfillOfficialContextWindow(p, 1_000_000)
-		}
-		return
-	}
-	entry := ProviderEntry{
-		Name:          "deepseek",
-		Kind:          "openai",
-		BaseURL:       "https://api.deepseek.com",
-		Models:        []string{"deepseek-v4-flash", "deepseek-v4-pro"},
-		Default:       "deepseek-v4-flash",
-		APIKeyEnv:     "DEEPSEEK_API_KEY",
-		BalanceURL:    "https://api.deepseek.com/user/balance",
-		ContextWindow: 1_000_000,
-		Prices:        deepSeekV4PricesForConfig(c),
-	}
-	if old, ok := c.Provider("deepseek-flash"); ok {
-		entry = officialProviderFromLegacy(entry, old)
-		entry.Prices = deepSeekV4PricesForConfig(c)
-		entry.Models = mergeModelLists([]string{"deepseek-v4-flash", "deepseek-v4-pro"}, old.ModelList())
-		entry.Default = firstKnownModel(entry.Default, entry.Models, "deepseek-v4-flash")
-	}
-	backfillOfficialContextWindow(&entry, 1_000_000)
-	c.Providers = append(c.Providers, entry)
-}
-
-func isOpenAIProviderKind(e *ProviderEntry) bool {
-	return e != nil && strings.EqualFold(strings.TrimSpace(e.Kind), "openai")
-}
-
-func mergeCuratedModelsIntoProvider(e *ProviderEntry, models []string, fallback string) {
-	// If the user has explicitly curated a model list (via Settings), respect
-	// that choice and do not merge additional curated models.
-	if len(e.Models) > 0 {
-		return
-	}
-	currentDefault := e.Default
-	if strings.TrimSpace(currentDefault) == "" {
-		currentDefault = e.Model
-	}
-	e.Models = mergeModelLists(models, e.ModelList())
-	e.Default = firstKnownModel(currentDefault, e.Models, fallback)
-}
-
-func backfillOfficialContextWindow(e *ProviderEntry, fallback int) {
-	if e != nil && e.ContextWindow <= 0 {
-		e.ContextWindow = fallback
-	}
-}
-
-func officialProviderFromLegacy(entry ProviderEntry, old *ProviderEntry) ProviderEntry {
-	entry.Kind = old.Kind
-	entry.BaseURL = old.BaseURL
-	entry.ModelsURL = old.ModelsURL
-	entry.APIKeyEnv = old.APIKeyEnv
-	entry.BalanceURL = old.BalanceURL
-	entry.ContextWindow = old.ContextWindow
-	entry.Price = old.Price
-	entry.Thinking = old.Thinking
-	entry.Effort = old.Effort
-	entry.ReasoningProtocol = old.ReasoningProtocol
-	entry.SupportedEfforts = append([]string(nil), old.SupportedEfforts...)
-	entry.DefaultEffort = old.DefaultEffort
-	entry.NoProxy = old.NoProxy
-	return entry
-}
-
-func mergeModelLists(primary, extra []string) []string {
-	seen := map[string]bool{}
-	out := make([]string, 0, len(primary)+len(extra))
-	for _, list := range [][]string{primary, extra} {
-		for _, model := range list {
-			model = strings.TrimSpace(model)
-			if model == "" || seen[model] {
-				continue
-			}
-			seen[model] = true
-			out = append(out, model)
-		}
-	}
-	return out
-}
-
-func firstKnownModel(current string, models []string, fallback string) string {
-	current = strings.TrimSpace(current)
-	for _, model := range models {
-		if model == current {
-			return current
-		}
-	}
-	for _, model := range models {
-		if model == fallback {
-			return fallback
-		}
-	}
-	if len(models) > 0 {
-		return models[0]
-	}
-	return ""
-}
-
-func retargetDesktopOfficialRefs(c *Config, access map[string]bool) {
-	c.DefaultModel = retargetDesktopOfficialRef(c.DefaultModel, access)
-	c.Agent.PlannerModel = retargetDesktopOfficialRef(c.Agent.PlannerModel, access)
-	c.Agent.SubagentModel = retargetDesktopOfficialRef(c.Agent.SubagentModel, access)
-	c.Agent.AutoPlanClassifier = retargetDesktopOfficialRef(c.Agent.AutoPlanClassifier, access)
-	for skill, ref := range c.Agent.SubagentModels {
-		c.Agent.SubagentModels[skill] = retargetDesktopOfficialRef(ref, access)
-	}
-}
-
-func retargetDesktopOfficialRef(ref string, access map[string]bool) string {
-	ref = strings.TrimSpace(ref)
-	if ref == "" {
-		return ""
-	}
-	provider, model, hasModel := strings.Cut(ref, "/")
-	switch provider {
-	case "deepseek-flash":
-		if !access["deepseek"] {
-			return ref
-		}
-		if !hasModel || strings.TrimSpace(model) == "" {
-			model = "deepseek-v4-flash"
-		}
-		return "deepseek/" + model
-	case "deepseek-pro":
-		if !access["deepseek"] {
-			return ref
-		}
-		if !hasModel || strings.TrimSpace(model) == "" {
-			model = "deepseek-v4-pro"
-		}
-		return "deepseek/" + model
-	default:
-		return ref
-	}
-}
-
-func userConfigPath() string {
-	dir := userConfigDir()
-	if dir == "" {
-		return ""
-	}
-	return filepath.Join(dir, "config.toml")
-}
-
-func userConfigDir() string {
-	return reasonixHomeDir()
-}
-
-func reasonixHomeDir() string {
-	if dir := cleanEnvDir("REASONIX_HOME"); dir != "" {
-		return dir
-	}
-	if runtime.GOOS != "windows" {
-		if home, err := os.UserHomeDir(); err == nil && home != "" {
-			return filepath.Join(home, ".reasonix")
-		}
-		return ""
-	}
-	dir := osUserConfigDir()
-	if dir == "" {
-		return ""
-	}
-	return filepath.Join(dir, "reasonix")
-}
-
-func userConfigLoadPath() string {
-	primary := userConfigPath()
-	if primary == "" {
-		return legacyUserConfigPath()
-	}
-	if _, err := os.Stat(primary); err == nil {
-		return primary
-	}
-	if legacy := legacyUserConfigPath(); legacy != "" {
-		if _, err := os.Stat(legacy); err == nil {
-			return legacy
-		}
-	}
-	for _, legacy := range legacyXDGConfigPaths() {
-		if legacy == "" || samePath(legacy, primary) {
-			continue
-		}
-		if _, err := os.Stat(legacy); err == nil {
-			return legacy
-		}
-	}
-	return primary
-}
-
-func legacyUserConfigPath() string {
-	dir := legacyOSSupportDir()
-	if dir == "" {
-		return ""
-	}
-	path := filepath.Join(dir, "config.toml")
-	if primary := userConfigPath(); primary != "" && samePath(path, primary) {
-		return ""
-	}
-	return path
-}
-
-func userConfigCandidatePaths() []string {
-	var paths []string
-	if p := userConfigPath(); p != "" {
-		paths = append(paths, p)
-	}
-	if p := legacyUserConfigPath(); p != "" {
-		paths = append(paths, p)
-	}
-	paths = append(paths, legacyXDGConfigPaths()...)
-	return paths
-}
-
-func legacyXDGConfigPaths() []string {
-	if runtime.GOOS == "windows" {
-		return nil
-	}
-	seen := map[string]bool{}
-	var paths []string
-	add := func(path string) {
-		if path == "" {
-			return
-		}
-		path = filepath.Clean(path)
-		if seen[path] {
-			return
-		}
-		seen[path] = true
-		paths = append(paths, path)
-	}
-	if dir := cleanEnvDir("XDG_CONFIG_HOME"); dir != "" {
-		add(filepath.Join(dir, "reasonix", "config.toml"))
-	}
-	if home, err := os.UserHomeDir(); err == nil && home != "" {
-		add(filepath.Join(home, ".config", "reasonix", "config.toml"))
-	}
-	return paths
-}
-
-func userSupportDir() string {
-	if dir := cleanEnvDir("REASONIX_STATE_HOME"); dir != "" {
-		return dir
-	}
-	return reasonixHomeDir()
-}
-
-func legacyOSSupportDir() string {
-	dir := osUserConfigDir()
-	if dir == "" {
-		return ""
-	}
-	path := filepath.Join(dir, "reasonix")
-	if current := reasonixHomeDir(); current != "" && samePath(path, current) {
-		return ""
-	}
-	return path
-}
-
-func userCacheDir() string {
-	if dir := cleanEnvDir("REASONIX_CACHE_HOME"); dir != "" {
-		return dir
-	}
-	dir, err := os.UserCacheDir()
-	if err != nil {
-		return ""
-	}
-	return filepath.Join(dir, "reasonix")
-}
-
-func osUserConfigDir() string {
-	dir, err := os.UserConfigDir()
-	if err != nil {
-		return ""
-	}
-	return dir
-}
-
-func cleanEnvDir(name string) string {
-	dir := strings.TrimSpace(os.Getenv(name))
-	if dir == "" {
-		return ""
-	}
-	dir = ExpandVars(dir)
-	if dir == "~" {
-		if home, err := os.UserHomeDir(); err == nil && home != "" {
-			dir = home
-		}
-	} else if strings.HasPrefix(dir, "~/") || strings.HasPrefix(dir, `~\`) {
-		if home, err := os.UserHomeDir(); err == nil && home != "" {
-			dir = filepath.Join(home, dir[2:])
-		}
-	}
-	if !filepath.IsAbs(dir) {
-		if abs, err := filepath.Abs(dir); err == nil {
-			dir = abs
-		}
-	}
-	return filepath.Clean(dir)
-}
-
-func samePath(a, b string) bool {
-	if a == "" || b == "" {
-		return false
-	}
-	aa, aerr := filepath.Abs(a)
-	bb, berr := filepath.Abs(b)
-	if aerr == nil {
-		a = aa
-	}
-	if berr == nil {
-		b = bb
-	}
-	return filepath.Clean(a) == filepath.Clean(b)
-}
-
-// userConfigDisplayPath is userConfigPath collapsed to a ~-relative form for
-// comments rendered into the user's own config.toml, so Windows users see the
-// real location instead of a hardcoded ~/.reasonix path.
-func userConfigDisplayPath() string {
-	p := userConfigPath()
-	if p == "" {
-		return "<os-config-dir>/reasonix/config.toml"
-	}
-	if home, err := os.UserHomeDir(); err == nil && home != "" {
-		if rel, err := filepath.Rel(home, p); err == nil && !strings.HasPrefix(rel, "..") {
-			return "~/" + filepath.ToSlash(rel)
-		}
-	}
-	return p
-}
-
-// UserConfigPath is the user-global config.toml. It lives under Reasonix home:
-// REASONIX_HOME/config.toml, then ~/.reasonix/config.toml on Unix-like systems,
-// or %AppData%/reasonix/config.toml on Windows. "" when the user config dir
-// can't be resolved.
-func UserConfigPath() string { return userConfigPath() }
-
-// LegacyUserConfigPath is the old OS app-support config.toml path when it
-// differs from UserConfigPath. It is read as a compatibility fallback when the
-// primary user config does not exist.
-func LegacyUserConfigPath() string { return legacyUserConfigPath() }
-
-// LegacyUserConfigPaths returns every known legacy user config path that differs
-// from the current v1.8.1 Reasonix-home config path.
-func LegacyUserConfigPaths() []string {
-	primary := userConfigPath()
-	var out []string
-	add := func(path string) {
-		if path == "" || samePath(path, primary) {
-			return
-		}
-		for _, existing := range out {
-			if samePath(existing, path) {
-				return
-			}
-		}
-		out = append(out, path)
-	}
-	add(legacyUserConfigPath())
-	for _, path := range legacyXDGConfigPaths() {
-		add(path)
-	}
-	return out
-}
-
-// ReasonixHomeDir is the current Reasonix home directory. It honors
-// REASONIX_HOME, then uses ~/.reasonix on macOS/Linux or %APPDATA%/reasonix on
-// Windows.
-func ReasonixHomeDir() string { return reasonixHomeDir() }
-
-// UserCredentialsPath is the reasonix-owned global .env file under Reasonix
-// home. It is the single source for provider credentials saved by Reasonix, so
-// stale shell, Windows, project, or home env vars cannot silently override keys
-// the user saved through setup or settings. "" when Reasonix home can't be
-// resolved.
-func UserCredentialsPath() string {
-	dir := userSupportDir()
-	if dir == "" {
-		return ""
-	}
-	return filepath.Join(dir, ".env")
-}
-
-// ArchiveDir is where compacted conversation history is archived for
-// traceability (one timestamped .jsonl per compaction). Empty if the user state
-// directory cannot be resolved, in which case archiving is skipped.
-func ArchiveDir() string {
-	dir := userSupportDir()
-	if dir == "" {
-		return ""
-	}
-	return filepath.Join(dir, "archive")
-}
-
-// SessionDir is where chat sessions are persisted (one .jsonl per session).
-// Used by `reasonix --continue` / `--resume` to find the recent ones. Empty
-// if the user state dir can't be resolved — sessions then aren't saved.
-func SessionDir() string {
-	dir := userSupportDir()
-	if dir == "" {
-		return ""
-	}
-	return filepath.Join(dir, "sessions")
-}
-
-// ProjectSessionDir is the per-workspace session directory the desktop sidebar
-// lists: <state root>/projects/<slug>/sessions. Empty when either the state root
-// or workspaceRoot doesn't resolve.
-func ProjectSessionDir(workspaceRoot string) string {
-	base := MemoryUserDir()
-	root := strings.TrimSpace(workspaceRoot)
-	if base == "" || root == "" {
-		return ""
-	}
-	if abs, err := filepath.Abs(root); err == nil {
-		root = abs
-	}
-	return filepath.Join(base, "projects", WorkspaceSlug(root), "sessions")
-}
-
-// WorkspaceSlug flattens an absolute workspace path into the directory name
-// used under <config root>/projects.
-func WorkspaceSlug(absPath string) string {
-	return strings.NewReplacer(string(os.PathSeparator), "-", "/", "-", "\\", "-", ":", "-").Replace(absPath)
-}
-
-// CacheDir is the per-user cache root for derived/regenerable artefacts: MCP
-// handshake snapshots, plugin startup-latency telemetry. Empty when the OS dir is
-// unavailable — callers must tolerate that (caching is best-effort).
-func CacheDir() string {
-	dir := userCacheDir()
-	if dir == "" {
-		return ""
-	}
-	return dir
-}
-
-// MemoryUserDir returns the reasonix user state root (…/reasonix), under which
-// the user-global REASONIX.md and the per-project auto-memory store live. Empty
-// when the user state dir can't be resolved, which disables user-scoped memory.
-func MemoryUserDir() string {
-	return userSupportDir()
-}
-
-// ConventionDirs are the parent directories scanned for agent assets (skills,
-// commands), in canonical-first order. .reasonix is ours; .agents / .agent /
-// .claude let users drop in assets authored for other agent tools without moving
-// files. Shared so skills (internal/skill) and commands (CommandDirs) discover
-// the same set. Note: hooks are NOT scanned across these — a .claude/settings.json
-// uses a different hook schema that can't be parsed as ours, so hooks stay in
-// .reasonix/settings.json (see internal/hook).
-var ConventionDirs = []string{".reasonix", ".agents", ".agent", ".claude"}
-
-// conventionSubdirsAsc joins sub under each ConventionDir of base, in ascending
-// priority (reverse of ConventionDirs) so the canonical .reasonix ends up the
-// highest-priority entry — command.Load lets a later directory win on a clash.
-func conventionSubdirsAsc(base, sub string) []string {
-	out := make([]string, 0, len(ConventionDirs))
-	for i := len(ConventionDirs) - 1; i >= 0; i-- {
-		out = append(out, filepath.Join(base, ConventionDirs[i], sub))
-	}
-	return out
-}
-
-// CommandDirs returns the directories scanned for custom slash commands, lowest
-// priority first, so a later (more specific) directory overrides an earlier one
-// on a name clash. Order: home-dir convention dirs (~/.claude/commands …
-// ~/.reasonix/commands), the Reasonix home commands dir, the legacy OS
-// app-support dir if different, then the project's
-// convention dirs (.claude/commands … .reasonix/commands). Scanning the .claude /
-// .agents / .agent dirs lets commands authored for other agent tools (same .md +
-// frontmatter format) work here unchanged.
-func CommandDirs() []string {
-	return CommandDirsForRoot(".")
-}
-
-// CommandDirsForRoot is like CommandDirs but resolves the project convention
-// dirs under root instead of the current working directory. Global dirs are
-// unchanged — they are always user-scoped.
-func CommandDirsForRoot(root string) []string {
-	root = resolveRoot(root)
-	var dirs []string
-	add := func(dir string) {
-		if dir == "" {
-			return
-		}
-		for _, existing := range dirs {
-			if samePath(existing, dir) {
-				return
-			}
-		}
-		dirs = append(dirs, dir)
-	}
-	if dir := legacyOSSupportDir(); dir != "" {
-		add(filepath.Join(dir, "commands"))
-	}
-	for _, legacy := range legacyXDGConfigPaths() {
-		add(filepath.Join(filepath.Dir(legacy), "commands"))
-	}
-	if home, err := os.UserHomeDir(); err == nil {
-		for _, dir := range conventionSubdirsAsc(home, "commands") {
-			add(dir)
-		}
-	}
-	if dir := userConfigDir(); dir != "" {
-		add(filepath.Join(dir, "commands"))
-	}
-	if dir := userSupportDir(); dir != "" && !samePath(dir, userConfigDir()) {
-		add(filepath.Join(dir, "commands"))
-	}
-	for _, dir := range conventionSubdirsAsc(root, "commands") {
-		add(dir)
-	}
-	return dirs
-}
-
-// SourcePath returns the highest-priority config file that exists, or "" if none.
-func SourcePath() string {
-	return SourcePathForRoot(".")
-}
-
-// SourcePathForRoot returns the highest-priority config file that exists under
-// root, or "" if none. Equivalent to SourcePath() when root is ".".
-func SourcePathForRoot(root string) string {
-	root = resolveRoot(root)
-	projectTOML := "reasonix.toml"
-	if root != "." {
-		projectTOML = filepath.Join(root, "reasonix.toml")
-	}
-	if _, err := os.Stat(projectTOML); err == nil {
-		return projectTOML
-	}
-	if uc := userConfigLoadPath(); uc != "" {
-		if _, err := os.Stat(uc); err == nil {
-			return uc
-		}
-	}
-	return ""
 }
 
 // WriteFile writes the configuration to path as annotated TOML. The write is
